animations and memory leaks
when applying the animation you'll see a hardcore memory leak that's still going on. I thought it had to do with the way it's drawn in shapes, but it's not so I'm sort of at a lost on what could cause memory to gain like that through an animation of rebuilding the shape in a new location over and over again.
Enable the animation in main.cpp at the world.updateSpecial()